pyproject.toml
[project]
name            = "uncle-bob"
version         = "1.0.0"
description     = "Curated GitHub Copilot skills, agents, prompts, and governance workflows."
readme          = "README.md"
requires-python = ">=3.10,<4"
dependencies    = [
  "typing-extensions>=4.12,<5",
]

[dependency-groups]
dev = [
  "pre-commit>=4.2,<5",
  "pyyaml>=6.0.2,<7",
  "ruff>=0.11,<0.12",
  "yamllint>=1.37,<2",
]

[tool.uv]
package = false
exclude-newer = "7 days"

[build-system]
requires      = ["hatchling>=1.27,<2"]
build-backend = "hatchling.build"
